Update README.md
I had everything ready to commit and running, but then when I hit to push, I ended up doing more since I got an error and basically lost all local files and it went back to Part 1 of Twitter app. I was in class with tech fellows Andres Aguilar and Liz Parra on 09/29, they can attest that they tried to help me. The gif above was taken before I committed and messed up the files. Both tech fellows tried to help me but in the end it seems as though it is lost :(